implemented realm again on the app and now I can write and receive data and update it to the UI

+++ TaylorSwiftQuiz/Questions/View Model/QuestionsViewModel.swift
@@ -68,7 +68,7 @@ class QuestionsViewModel {
         return UserDataModel.shared.newPlayer?.currentQuestion ?? 0
     }
     
-    public func getCurrentPoints() -> Int16 {
+    public func getCurrentPoints() -> Int32 {
         return UserDataModel.shared.newPlayer?.points ?? 0
     }
